+++ empire/test/test_agents.py
@@ -1,11 +1,66 @@
+import base64
 import logging
+import struct
 import time
+import zlib
 from datetime import UTC, datetime, timedelta
+from pathlib import Path
 
 import pytest
 from sqlalchemy.exc import IntegrityError
 
+from empire.server.common.empire import MainMenu
+from empire.server.utils.string_util import is_valid_session_id
+
 log = logging.getLogger(__name__)
+
+
+# Copied from the agent.py file for Python agent
+class compress:
+    """
+    Base clase for init of the package. This will handle
+    the initial object creation for conducting basic functions.
+    """
+
+    CRC_HSIZE = 4
+    COMP_RATIO = 9
+
+    def __init__(self, verbose=False):
+        """
+        Populates init.
+        """
+        pass
+
+    def comp_data(self, data, cvalue=COMP_RATIO):
+        """
+        Takes in a string and computes
+        the comp obj.
+        data = string wanting compression
+        cvalue = 0-9 comp value (default 6)
+        """
+        cdata = zlib.compress(data, cvalue)
+        return cdata
+
+    def crc32_data(self, data):
+        """
+        Takes in a string and computes crc32 value.
+        data = string before compression
+        returns:
+        HEX bytes of data
+        """
+        crc = zlib.crc32(data) & 0xFFFFFFFF
+        return crc
+
+    def build_header(self, data, crc):
+        """
+        Takes comp data, org crc32 value,
+        and adds self header.
+        data =  comp data
+        crc = crc32 value
+        """
+        header = struct.pack("!I", crc)
+        built_data = header + data
+        return built_data
 
 
 @pytest.fixture(scope="function", autouse=True)
@@ -220,3 +275,141 @@ def test_can_ignore_duplicate_checkins(session_local, models, agent, main):
         checkin_count = len(db_agent.checkins.all())
 
         assert checkin_count == prev_checkin_count + 1
+
+
+def test_update_dir_list(session_local, models, agent, main: MainMenu):
+    with session_local.begin() as db:
+        message = {
+            "directory_path": "C:\\",
+            "directory_name": "C:\\",
+            "items": [
+                {
+                    "path": "C:\\Users\\vinnybod\\Desktop\\test.txt",
+                    "name": "test.txt",
+                    "is_file": True,
+                },
+                {
+                    "path": "C:\\Users\\vinnybod\\Desktop\\test2.txt",
+                    "name": "test2.txt",
+                    "is_file": True,
+                },
+            ],
+        }
+        main.agentcommsv2._update_dir_list(db, agent, message)
+
+        file, _ = main.agentfilesv2.get_file_by_path(
+            db, agent, "C:\\Users\\vinnybod\\Desktop\\test.txt"
+        )
+
+        assert file.path == "C:\\Users\\vinnybod\\Desktop\\test.txt"
+        assert file.name == "test.txt"
+        assert file.is_file is True
+        assert file.parent_id is not None
+
+        db.query(models.AgentFile).delete()
+
+
+def test_update_dir_list_with_existing_joined_file(
+    session_local, models, agent, main: MainMenu, empire_config
+):
+    with session_local.begin() as db:
+        message = {
+            "directory_path": "C:\\",
+            "directory_name": "C:\\",
+            "items": [
+                {
+                    "path": "C:\\Users\\vinnybod\\Desktop\\test.txt",
+                    "name": "test.txt",
+                    "is_file": True,
+                },
+                {
+                    "path": "C:\\Users\\vinnybod\\Desktop\\test2.txt",
+                    "name": "test2.txt",
+                    "is_file": True,
+                },
+            ],
+        }
+        main.agentcommsv2._update_dir_list(db, agent, message)
+
+        file, _ = main.agentfilesv2.get_file_by_path(
+            db, agent, "C:\\Users\\vinnybod\\Desktop\\test.txt"
+        )
+
+        download_path = Path("empire/test/avatar.png")
+        file.downloads.append(
+            models.Download(
+                location=str(download_path.absolute()),
+                filename=download_path.name,
+                size=download_path.stat().st_size,
+            )
+        )
+
+        # This previously raised a Foreign Key Constraint error, but should succeed now.
+        main.agentcommsv2._update_dir_list(db, agent, message)
+
+        file2, _ = main.agentfilesv2.get_file_by_path(
+            db, agent, "C:\\Users\\vinnybod\\Desktop\\test.txt"
+        )
+
+        if empire_config.database.use != "sqlite":
+            # sqlite reuses ids and apparently doesn't cascade the delete to the
+            # association table. This can result in files being linked to the wrong
+            # download after refreshing a directory for sqlite.
+            assert file.id != file2.id
+            assert len(file2.downloads) == 0
+
+        assert file.name == file2.name
+
+        db.query(models.AgentFile).delete()
+
+
+@pytest.mark.parametrize(
+    "session_id,expected",
+    [
+        ("ABCDEFGH", True),
+        ("12345678", True),
+        ("ABCDEF1H", True),
+        ("A1B2C3D4", True),
+        ("ABCDEFG", False),
+        ("ABCDEFGHI", False),
+        ("ABCD_EFG", False),
+        ("       ", False),
+        ("", False),
+        (12345678, False),
+        (None, False),
+        ("./../../", False),
+    ],
+)
+def test_is_valid_session_id(session_id, expected):
+    assert (
+        is_valid_session_id(session_id) == expected
+    ), f"Test failed for session_id: {session_id}"
+
+
+def test_skywalker_exploit_protection(caplog, agent, db, main: MainMenu):
+    # Malicious file path attempting directory traversal
+    malicious_directory = (
+        main.installPath + r"/downloads/..\\..\\..\\..\\..\\etc\\cron.d\\evil"
+    )
+    encodedPart = b"test"
+    c = compress()
+    start_crc32 = c.crc32_data(encodedPart)
+    comp_data = c.comp_data(encodedPart)
+    encodedPart = c.build_header(comp_data, start_crc32)
+    encodedPart = base64.b64encode(encodedPart).decode("UTF-8")
+
+    malicious_data = "|".join(
+        [
+            "0",
+            malicious_directory,
+            "6",
+            encodedPart,
+        ]
+    )
+
+    main.agentcommsv2._process_agent_packet(
+        db, agent, "TASK_DOWNLOAD", "1", malicious_data
+    )
+
+    expected_message_part = "attempted skywalker exploit!"
+    assert any(expected_message_part in message for message in caplog.messages)